refactor(goto): move client hints out of the screen PR
UA Client Hints now ship in their own PR. This branch keeps only the screen emulation, in packages/goto/src/screen.js, with the goto e2e and unit tests that cover it.
